Move Consumer vote processing loops to VoteWorker
